refactor: clean up imports by removing unused AuthStatus from UserContext
