+++ apps/chat/src/pages/auth/Login.spec.tsx
@@ -2,10 +2,10 @@ import { render, screen } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { AuthI18nKeys } from '../../constants/translation-keys';
-import { AuthStatus } from '../../context/auth/UserContext';
 import * as UserContextModule from '../../context/auth/UserContext';
 import * as useAuthRedirectModule from '../../hooks/auth/useAuthRedirect';
 import * as authApi from '../../server-api/auth.api';
+import { AuthStatus } from '../../types/auth-status';
 import LoginPage from './Login';
 
 vi.mock('../../context/auth/UserContext');
